requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
werkzeug==3.0.1
gunicorn==21.2.0
SQLAlchemy==2.0.23
Jinja2==3.1.2
itsdangerous==2.1.2
greenlet==3.0.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
typing_extensions==4.8.0
urllib3==2.1.0
Flask-Migrate==4.0.5

